Update old components
